requirements.txt
lightning-flash
lightning-flash[image]
gradio